+++ backend/src/services/posts.rs
@@ -99,7 +99,7 @@ pub async fn create_post(
         username: Some(user.username),
         avatar_url: user.avatar_url,
         email: Some(user.email),
-        reply_count: post.reply_count,
+        reply_count: post.reply_count as i64,
         last_reply_at: post.last_reply_at,
         files: vec![],
         reactions: vec![],
